requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
docxtpl==0.16.8
python-docx==1.1.2
SQLAlchemy[asyncio]==2.0.30
aiomysql==0.2.0
aiofiles==23.2.1
httpx==0.27.0
jinja2==3.1.4
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery==5.4.0
redis==5.0.4
openpyxl==3.1.3
tenacity==8.3.0
structlog==24.2.0
